[Rename] api 통신 폴더 변경 (#147)
* docs: GitHub 협업 및 브랜치 관리 정책 문서 추가

* Feat: focus, tab막기

* Feat: api호출 customQuery 로 변경

* Feat: 에러 로딩 처리 목표, 노트, 할일

* Chore: eslint if 중첩 오류

* Comment: 주석삭제

* Refactor: 피드백 관련 수정

* Feat: toast 추가

* Chore: api 호출 함수 코드 변경

* Design: 노트모아보기 스타일 추가

* Fix: 화살표함수수정

* Rename: 파일변경

* Fix: import 수정

* Fix: 화살표 함수

* Refactor: type 정의

* Feat: 노트작성 제목과내용입력 시 작성완료버튼 활성화

* Refactor: 노트작성하기 버튼활성화 코드 수정

* Chore: 변수명 수정

* Rename: api 통신 폴더 변경

* Rename: api 통신 파일 변경

---------
