+++ src/app/goals/[goalId]/page.tsx
@@ -7,8 +7,6 @@ import {useEffect, useState} from 'react'
 import {useQueryClient} from '@tanstack/react-query'
 import axios from 'axios'
 
-import {goalDataApi, goalDeleteApi, goalUpdateApi} from '@/app/api/goal-api'
-import {todoDeleteApi, todoUpdateApi} from '@/app/api/todo-api'
 import LoadingSpinner from '@/components/common/loading-spinner'
 import AddTodoModal from '@/components/common/modal/add-todo-modal'
 import TwoButtonModal from '@/components/common/modal/two-buttom-modal'
@@ -20,6 +18,8 @@ import {useInfiniteScrollQuery} from '@/hooks/use-infinite-scroll'
 import useModal from '@/hooks/use-modal'
 import useToast from '@/hooks/use-toast'
 import {get} from '@/lib/api'
+import {goalDataApi, goalDeleteApi, goalUpdateApi} from '@/lib/goals/api'
+import {todoDeleteApi, todoUpdateApi} from '@/lib/todos/api'
 import {useModalStore} from '@/store/use-modal-store'
 
 import type {Goal} from '@/types/goals'
