+++ src/app/notes/write/page.tsx
@@ -5,13 +5,13 @@ import {useEffect} from 'react'
 
 import axios from 'axios'
 
-import {goalDataApi} from '@/app/api/goal-api'
-import {todoDataApi} from '@/app/api/todo-api'
 import LoadingSpinner from '@/components/common/loading-spinner'
 import NoteEditCompo from '@/components/notes/edit'
 import NoteWriteCompo from '@/components/notes/write'
 import {useCustomQuery} from '@/hooks/use-custom-query'
 import useToast from '@/hooks/use-toast'
+import {goalDataApi} from '@/lib/goals/api'
+import {todoDataApi} from '@/lib/todos/api'
 
 import type {Goal} from '@/types/goals'
