+++ src/components/notes/write.tsx
@@ -8,13 +8,13 @@ import axios from 'axios'
 
 import 'react-toastify/dist/ReactToastify.css'
 
-import {noteRegApi} from '@/app/api/note-api'
 import TwoButtonModal from '@/components/common/modal/two-buttom-modal'
 import MarkdownEditor from '@/components/markdown-editor/markdown-editor'
 import ButtonStyle from '@/components/style/button-style'
 import {useCustomMutation} from '@/hooks/use-custom-mutation'
 import useModal from '@/hooks/use-modal'
 import useToast from '@/hooks/use-toast'
+import {noteRegApi} from '@/lib/notes/api'
 import {useModalStore} from '@/store/use-modal-store'
 import {getTextFromHtml} from '@/utils/text-from-html'
